+++ DelvUI/Helpers/ClipRectsHelper.cs
@@ -107,15 +107,21 @@ namespace DelvUI.Helpers
                     float margin = 5 * addon->Scale;
                     float bottomMargin = 13 * addon->Scale;
 
-                    ClipRect clipRect = new ClipRect(
-                        new Vector2(addon->X + margin, addon->Y + margin),
-                        new Vector2(
-                            addon->X + addon->WindowNode->AtkResNode.Width * addon->Scale - margin,
-                            addon->Y + addon->WindowNode->AtkResNode.Height * addon->Scale - bottomMargin
-                        )
+                    Vector2 pos = new Vector2(addon->X + margin, addon->Y + margin);
+                    Vector2 size = new Vector2(
+                        addon->WindowNode->AtkResNode.Width * addon->Scale - margin,
+                        addon->WindowNode->AtkResNode.Height * addon->Scale - bottomMargin
                     );
 
+                    // special case for duty finder
+                    if (name == "ContentsFinder")
+                    {
+                        size.X += size.X + (16 * addon->Scale);
+                        size.Y += (30 * addon->Scale);
+                    }
+
                     // just in case this causes weird issues / crashes (doubt it though...)
+                    ClipRect clipRect = new ClipRect(pos, pos + size);
                     if (clipRect.Max.X < clipRect.Min.X || clipRect.Max.Y < clipRect.Min.Y)
                     {
                         continue;